+++ kermit/k95/kui/kui.cxx
@@ -390,8 +390,13 @@ void Kui::setProperty( int propid, intptr_t param1, intptr_t param2 )
 }
 /*------------------------------------------------------------------------
 ------------------------------------------------------------------------*/
-void Kui::getProperty( int propid, long param1, long param2 )
+int Kui::getProperty( int propid, void* out )
 {
+    switch (propid) {
+        case KUI_GUI_TOOLBAR_VIS: {
+            return terminal->getToolbarVisible();
+        }
+    }
 }
 
 /*------------------------------------------------------------------------